inheritance2/Blockchain_backend/scraping/requirements.txt
# PostgreSQL dependencies for scraping
playwright==1.48.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
